imap_prefix_mail/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.0
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
imapclient==2.3.1
python-dotenv==1.0.1
jinja2==3.1.4
